Create Movement Manager
Creates a Movement manager to handle movement instead of the
Game Manager. This allow sus to separate our concerns more easily
