+++ SwipeRogueUnity/Assets/Scripts/MovementManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 580c14e952693481682b70836ea21e44
+timeCreated: 1488777925
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
